backend/pyproject.toml
[project]
name = "cricket-batting-analysis"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.12",
    "arq>=0.26",
    "redis>=5.0",
    "anthropic>=0.40",
    "opencv-python-headless>=4.10",
    "numpy>=1.26",
    "mediapipe>=0.10",
    "python-dotenv>=1.0",
    "aiofiles>=23.2",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23", "httpx>=0.27"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
